refactor(web): move deployments to @safe-global/utils
