+++ apps/web/src/components/common/NetworkSelector/NetworkMultiSelector.tsx
@@ -10,7 +10,7 @@ import { useRouter } from 'next/router'
 import { getNetworkLink } from '.'
 import { SetNameStepFields } from '@/components/new-safe/create/steps/SetNameStep'
 import { getSafeSingletonDeployments, getSafeToL2SetupDeployments } from '@safe-global/safe-deployments'
-import { hasCanonicalDeployment } from '@/services/contracts/deployments'
+import { hasCanonicalDeployment } from '@safe-global/utils/services/contracts/deployments'
 import { hasMultiChainCreationFeatures } from '@/features/multichain/utils/utils'
 import { getLatestSafeVersion } from '@safe-global/utils/utils/chains'
