+++ apps/web/src/components/transactions/TxDetails/TxData/MigrationToL2TxData/index.tsx
@@ -3,7 +3,7 @@ import useAsync from '@/hooks/useAsync'
 import { useCurrentChain } from '@/hooks/useChains'
 import useSafeInfo from '@/hooks/useSafeInfo'
 import { useWeb3ReadOnly } from '@/hooks/wallets/web3'
-import { getMultiSendContractDeployment } from '@/services/contracts/deployments'
+import { getMultiSendContractDeployment } from '@safe-global/utils/services/contracts/deployments'
 import { createTx } from '@/services/tx/tx-sender/create'
 import { Safe__factory } from '@/types/contracts'
 import { type TransactionDetails } from '@safe-global/safe-gateway-typescript-sdk'
