+++ apps/web/src/services/security/modules/DelegateCallModule/index.ts
@@ -1,5 +1,5 @@
 import { OperationType } from '@safe-global/safe-core-sdk-types'
-import { getMultiSendCallOnlyContractDeployment } from '@/services/contracts/deployments'
+import { getMultiSendCallOnlyContractDeployment } from '@safe-global/utils/services/contracts/deployments'
 import type { SafeTransaction } from '@safe-global/safe-core-sdk-types'
 import type { ChainInfo, SafeInfo } from '@safe-global/safe-gateway-typescript-sdk'
